modules 2.6.24-hhtech

+++ initramfs/compress.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 #use "apt-get uboot-mkimage" for the mkimage tool
 
@@ -6,13 +6,19 @@
 SDPART=/media/2B3F-7268
 
 #kernel to use
-#ZIMAGE=./zimage
-ZIMAGE=./zImageSD
+ZIMAGE=./zimage
+#ZIMAGE=./zImageSD
+#ZIMAGE=./zImageMer
 #ZIMAGE=./zImage2631q
 #ZIMAGE=./zImage34
 #ZIMAGE=./zImage2635
 
 rm initrd.igz
+
+#repair lost rights in git
+./prepare-dev.sh
+chmod +s content/bin/busybox
+chmod +x content/sbin/init
 
 (cd content; find . -print | cpio -o --format=newc --owner=0) | gzip -c -9 > tmp.gz
 
@@ -20,7 +26,7 @@ mkimage -d tmp.gz -A ARM -O Linux -C gzip -T RAMDisk -n "SmartQ initramfs" initr
 
 ../fw-utils/mkSmartQ5 no-qi.nb0 no-u-boot.bin $ZIMAGE initrd.igz
 
-rm $SDPART/SmartQ5
+[ -f $SDPART/SmartQ5 ] && rm $SDPART/SmartQ5
 cp SmartQ5 $SDPART/
 
 umount $SDPART
